Move documentStateFor() to services.js (#10340)
* move documentStateFor() to services.js

* move install to runtime.js

* fix lint and types

* fix tests
