+++ extensions/amp-analytics/0.1/test/test-visibility-manager.js
@@ -23,7 +23,7 @@ import {
   VisibilityManagerForEmbed,
 } from '../visibility-manager';
 import {VisibilityState} from '../../../../src/visibility-state';
-import {documentStateFor} from '../../../../src/service/document-state';
+import {documentStateFor} from '../../../../src/services';
 import {layoutRectLtwh, rectIntersection} from '../../../../src/layout-rect';
 
 class IntersectionObserverStub {
