+++ extensions/amp-dynamic-css-classes/0.1/test/test-runtime-classes.js
@@ -16,6 +16,9 @@
 
 import {ampdocServiceFor, viewerForDoc} from '../../../../src/services';
 import {installDocService} from '../../../../src/service/ampdoc-impl';
+import {
+  installDocumentStateService,
+} from '../../../../src/service/document-state';
 import {installPlatformService} from '../../../../src/service/platform-impl';
 import {installViewerServiceForDoc} from '../../../../src/service/viewer-impl';
 import {vsyncForTesting} from '../../../../src/service/vsync-impl';
@@ -62,6 +65,7 @@ describe('dynamic classes are inserted at runtime', () => {
 
   function setup(embeded, userAgent, referrer) {
     installDocService(mockWin, /* isSingleDoc */ true);
+    installDocumentStateService(mockWin);
     const ampdocService = ampdocServiceFor(mockWin);
     const ampdoc = ampdocService.getAmpDoc();
     installPlatformService(mockWin);
