+++ src/service/document-state.js
@@ -17,7 +17,7 @@
 import {Observable} from '../observable';
 import {getVendorJsPropertyName} from '../style';
 import {waitForChild} from '../dom';
-import {registerServiceBuilder, getService} from '../service';
+import {registerServiceBuilder} from '../service';
 
 /**
  */
@@ -150,13 +150,4 @@ export class DocumentState {
  */
 export function installDocumentStateService(window) {
   registerServiceBuilder(window, 'documentState', DocumentState);
-}
-
-/**
- * @param {!Window} window
- * @return {!DocumentState}
- */
-export function documentStateFor(window) {
-  installDocumentStateService(window);
-  return getService(window, 'documentState');
 }
