+++ test/functional/test-custom-element.js
@@ -19,6 +19,7 @@ import {AmpEvents} from '../../src/amp-events';
 import {BaseElement} from '../../src/base-element';
 import {ElementStub, setLoadingCheckForTests} from '../../src/element-stub';
 import {LOADING_ELEMENTS_, Layout} from '../../src/layout';
+import {installDocumentStateService} from '../../src/service/document-state';
 import {installResourcesServiceForDoc} from '../../src/service/resources-impl';
 import {poll} from '../../testing/iframe';
 import {ResourceState} from '../../src/service/resource';
@@ -1884,6 +1885,8 @@ describes.realWin('CustomElement Overflow Element', {amp: true}, env => {
         ampExtendedElements: {},
       };
       doc.defaultView = win;
+
+      installDocumentStateService(win);
     });
 
     afterEach(() => {
