+++ test/functional/test-viewer.js
@@ -18,6 +18,7 @@ import {Viewer} from '../../src/service/viewer-impl';
 import {ampdocServiceFor} from '../../src/services';
 import {dev} from '../../src/log';
 import {installDocService} from '../../src/service/ampdoc-impl';
+import {installDocumentStateService} from '../../src/service/document-state';
 import {installPlatformService} from '../../src/service/platform-impl';
 import {installTimerService} from '../../src/service/timer-impl';
 import {parseUrl, removeFragment} from '../../src/url';
@@ -83,6 +84,7 @@ describe('Viewer', () => {
       windowApi.location.href = url;
     });
     installDocService(windowApi, /* isSingleDoc */ true);
+    installDocumentStateService(windowApi);
     ampdoc = ampdocServiceFor(windowApi).getAmpDoc();
     installPlatformService(windowApi);
     installTimerService(windowApi);
