+++ test/integration/test-visibility-states.js
@@ -14,9 +14,11 @@
  * limitations under the License.
  */
 
-import {viewerPromiseForDoc} from '../../src/services';
-import {documentStateFor} from '../../src/service/document-state';
-import {resourcesForDoc} from '../../src/services';
+import {
+  documentStateFor,
+  resourcesForDoc,
+  viewerPromiseForDoc,
+} from '../../src/services';
 import {VisibilityState} from '../../src/visibility-state';
 import {getVendorJsPropertyName} from '../../src/style';
 import {whenUpgradedToCustomElement} from '../../src/dom';
